requirements.txt
Django==1.6.8
SQLAlchemy==1.0.14
South==1.0.1
argparse==1.2.1
django-tastypie==0.12.1
psycopg2==2.6.1
python-dateutil==2.5.3
python-mimeparse==1.5.3
six==1.10.0
unicodecsv==0.14.1
wsgiref==0.1.2
django_pg==1.4.2